refactor: move from ValueObjects dir as these are not value objects
